+++ cpp/oneapi/dal/backend/primitives/optimizers/cg_solver.hpp
@@ -31,9 +31,9 @@ sycl::event cg_solve(sycl::queue& queue,
                      ndview<Float, 1>& residual,
                      ndview<Float, 1>& conj_vector,
                      ndview<Float, 1>& buffer,
-                     const Float tol = 1e-5,
-                     const Float atol = -1,
-                     const std::int64_t maxiter = 100,
+                     Float tol = 1e-5,
+                     Float atol = -1,
+                     std::int64_t maxiter = 100,
                      const event_vector& deps = {});
 
 } // namespace oneapi::dal::backend::primitives
