+++ cpp/oneapi/dal/backend/primitives/optimizers/test/fixture.hpp
@@ -17,12 +17,81 @@
 #pragma once
 
 #include <type_traits>
+#include "oneapi/dal/backend/primitives/optimizers/common.hpp"
 #include "oneapi/dal/backend/primitives/ndarray.hpp"
 #include "oneapi/dal/test/engine/common.hpp"
 #include "oneapi/dal/test/engine/fixtures.hpp"
 #include "oneapi/dal/backend/primitives/rng/rng_engine.hpp"
 
 namespace oneapi::dal::backend::primitives::test {
+
+// f(x) = 1/2 x^t A x - b^t x
+// df / dx = Ax - b
+// df / d^2x = A
+template <typename Float>
+class QuadraticFunction : public BaseFunction<Float> {
+public:
+    QuadraticFunction(sycl::queue& q, const ndview<Float, 2>& A, const ndview<Float, 1>& b)
+            : q_(q),
+              n_(A.get_dimension(0)),
+              A_(A),
+              b_(b),
+              hessp_(q, A) {
+        ONEDAL_ASSERT(A.get_dimension(1) == n_);
+        ONEDAL_ASSERT(b.get_dimension(0) == n_);
+        gradient_ = ndarray<Float, 1>::empty(q, { n_ }, sycl::usm::alloc::device);
+        tmp_ = ndarray<Float, 1>::empty(q, { 1 }, sycl::usm::alloc::device);
+    }
+    ~QuadraticFunction() {}
+    Float get_value() override {
+        return value_;
+    }
+
+    ndview<Float, 1>& get_gradient() override {
+        return gradient_;
+    }
+
+    BaseMatrixOperator<Float>& get_hessian_product() override {
+        return hessp_;
+    }
+
+    sycl::event update_x(const ndview<Float, 1>& x,
+                         bool need_hessp = false,
+                         const event_vector& deps = {}) override {
+        constexpr Float zero(0), one(1);
+        auto fill_gradient_event = fill<Float>(q_, gradient_, zero, deps);
+        auto fill_value_event = fill<Float>(q_, tmp_, zero, deps);
+        auto gemv_event = gemv(q_, A_, x, gradient_, one, zero, { fill_gradient_event }); // Ax
+        Float tmp_host = 0;
+        auto xtax_event = dot_product(q_,
+                                      gradient_,
+                                      x,
+                                      tmp_.get_mutable_data(),
+                                      &tmp_host,
+                                      { gemv_event }); //x^tAx
+        auto btx_event =
+            dot_product(q_, b_, x, tmp_.get_mutable_data(), &value_, { xtax_event }); // b^t x
+        const auto kernel_minus = [=](const Float a, const Float b) -> Float {
+            return a - b;
+        };
+        auto bias_event = element_wise(q_, kernel_minus, gradient_, b_, gradient_, { xtax_event });
+
+        btx_event.wait_and_throw();
+        value_ = -value_ + tmp_host / 2; // 1/2 x^t A x - b^t x
+
+        return bias_event;
+    }
+
+private:
+    sycl::queue q_;
+    const std::int64_t n_;
+    const ndview<Float, 2> A_;
+    const ndview<Float, 1> b_;
+    Float value_;
+    ndarray<Float, 1> tmp_;
+    ndarray<Float, 1> gradient_;
+    LinearMatrixOperator<Float> hessp_;
+};
 
 template <typename Float>
 void check_val(const Float real, const Float expected, const Float rtol, const Float atol) {
